get chefmenu by id working
get chefmenu by id working

+++ src/main/java/ethniconnect_backend/ChefCreateMenu/ChefMenu.java
@@ -28,7 +28,7 @@ public class ChefMenu {
     @Column(columnDefinition = "MEDIUMBLOB")
     private String menu_item_image;
     private String item_name;
-    private int menu_item_price;
+    private double menu_item_price;
     private String item_ingredients;
     private String item_intresting_facts;
     @Enumerated(EnumType.STRING)
